Renamed Hello Run Configuration to Main Added sqlite-jdbc dependency to POM Refactored java files into the appropriate Model & Controller packages
