+++ .idea/runConfigurations/Main.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Hello" type="Application" factoryName="Application">
+  <configuration default="false" name="Main" type="Application" factoryName="Application">
     <option name="MAIN_CLASS_NAME" value="com.example.napolinights.Main" />
     <module name="Napoli-Nights" />
     <method v="2">
