+++ src/main/java/com/example/napolinights/Main.java
@@ -1,11 +1,7 @@
 package com.example.napolinights;
 
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
+import com.example.napolinights.model.*;
 
-import java.io.IOException;
 import java.sql.Connection;
 import java.util.List;
