+++ src/main/java/com/example/napolinights/model/Category.java
@@ -1,4 +1,4 @@
-package com.example.napolinights;
+package com.example.napolinights.model;
 
 public enum Category {
     ENTREE,
