+++ src/main/java/module-info.java
@@ -6,4 +6,8 @@ module com.example.napolinights {
 
     opens com.example.napolinights to javafx.fxml;
     exports com.example.napolinights;
+    exports com.example.napolinights.model;
+    opens com.example.napolinights.model to javafx.fxml;
+    exports com.example.napolinights.controller;
+    opens com.example.napolinights.controller to javafx.fxml;
 }
